pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civitscraper"
version = "0.1.0"
description = "A tool for fetching and managing CivitAI model metadata"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "AGPL-3.0"}
dependencies = [
    "requests>=2.25.0",
    "pyyaml>=5.4.0",
    "jinja2>=3.0.0",
    "blake3>=0.3.0",
]

[project.scripts]
civitscraper = "civitscraper.cli:main"

[tool.setuptools]
packages = {find = {include = ["civitscraper", "civitscraper.*"]}}
include-package-data = true

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
